+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="$USER_HOME$/.android/avd/Pixel_3a_API_TiramisuPrivacySandbox.avd" />
+            <value value="C:\Users\hanti\.android\avd\Pixel_3a_API_TiramisuPrivacySandbox.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-11-15T08:52:55.915576Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-11-22T06:28:33.457556100Z" />
   </component>
 </project>
